Amend Storage to reject dates past system date
Storage adds entries to the FinancialList directly
via FinancialList.addEntry(). This bypasses the check for future
dates in AddExpenseCommand/AddIncomeCommand.

Let's modify Storage to add entries to FinancialList using
AddExpenseCommand and AddIncomeCommand to ensure that entries
with dates past the system date are ignored.

Additionally, Storage was also modified such that when
it processes an errenous entry, it skips the single entry instead
of returning null for the FinancialList.